Refactor(WalletConnect): centralize session state in context
