+++ apps/web/src/features/walletconnect/__tests__/WalletConnectContext.test.tsx
@@ -8,7 +8,7 @@ import type { SessionTypes } from '@walletconnect/types'
 import { act, fireEvent, render, waitFor } from '@/tests/test-utils'
 import { WalletConnectContext } from '../WalletConnectContext'
 import WalletConnectWallet from '../services/WalletConnectWallet'
-import { WalletConnectProvider } from '../components/WalletConnectProvider'
+import { WalletConnectProvider } from '../WalletConnectContext'
 import { safeInfoSlice } from '@/store/safeInfoSlice'
 import { useAppDispatch } from '@/store'
 import * as useSafeWalletProvider from '@/services/safe-wallet-provider/useSafeWalletProvider'
